+++ src/onegov.election_day/onegov/election_day/views/manage/data_sources.py
@@ -127,6 +127,7 @@ def view_data_source_items(self, request):
         'layout': ManageDataSourceItemsLayout(self, request),
         'title': _("Manage"),
         'items': self.batch,
+        'source': self.source,
         'item_name': self.source.label,
         'source': self.source,
         'new_item': request.link(self, 'new-item')